fix: Update the Name Squatting changes to follow project conventions
