+++ server/src/main/java/org/eclipse/openvsx/repositories/RepositoryService.java
@@ -1483,7 +1483,7 @@ public class RepositoryService {
             @Nullable String name,
             @Nullable LocalDateTime detectedFrom,
             @Nullable LocalDateTime detectedTo,
-            org.eclipse.openvsx.admin.NameSquattingAPI.@Nullable ExtensionStateFilter stateFilter,
+            org.eclipse.openvsx.admin.AdminService.@Nullable ExtensionStateFilter stateFilter,
             boolean ascending,
             int limit,
             int offset
@@ -1515,7 +1515,7 @@ public class RepositoryService {
             @Nullable String name,
             @Nullable LocalDateTime detectedFrom,
             @Nullable LocalDateTime detectedTo,
-            org.eclipse.openvsx.admin.NameSquattingAPI.@Nullable ExtensionStateFilter stateFilter
+            org.eclipse.openvsx.admin.AdminService.@Nullable ExtensionStateFilter stateFilter
     ) {
         var applyStateFilter = stateFilter != null && stateFilter.hasFilter();
         return extensionValidationFailureRepo.countFlaggedExtensions(
